streamlit_app/requirements.txt
streamlit==1.36.0
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.2